TestEngine: Viewport support: clear ImGuiBackendFlags_HasMouseHoveredViewport flag when tests run.
Purpose of this flag is to allow third party windows to appear in-between imgui windows. However hovered viewport reported by the backend may not be valid during testing, because simulated mouse position would not match position of mouse cursor and wrong window would be reported. Test engine could override g.IO.MouseHoveredViewport value, but to do that reliably we would have to copy behavior of FindHoveredWindow() to perform this search. It is a complicated solution for something test engine does not really need, therefore this flag is disabled during test execution.